BVH half split done
Now do SAH

+++ AdvancedTechnologiesProject2/Src/ComparePrimitives.cpp
@@ -9,8 +9,8 @@ ComparePrimitives::ComparePrimitives(Axis axis)
 
 bool ComparePrimitives::operator()(std::shared_ptr<Geometry> a, std::shared_ptr<Geometry> b)
 {
-	float ca = a->getPos()[m_axis];
-	float cb = b->getPos()[m_axis];
+	float ca = a->getPos()[static_cast<int>(m_axis)];
+	float cb = b->getPos()[static_cast<int>(m_axis)];
 								  
 	return ca < cb;
 }
